chore: parametrize import test
